refactor: Split controllers into separate files
