+++ router/router.go
@@ -1,0 +1,22 @@
+package router
+
+import (
+	"go-api-tut/controllers"
+
+	"github.com/gin-contrib/cors"
+	"github.com/gin-gonic/gin"
+)
+
+func SetupRouter() *gin.Engine {
+	r := gin.Default()
+
+	r.Use(cors.Default())
+
+	r.GET("/users", controllers.GetUsers)
+	r.POST("/users", controllers.CreateUser)
+	r.GET("/users/:id", controllers.GetUser)
+	r.PUT("/users/:id", controllers.UpdateUser)
+	r.DELETE("/users/:id", controllers.DeleteUser)
+
+	return r
+}
